* changed tasks to be controlled by task controller, this should fix renaming tasks while the timer is running * migrated a lot of code to objc2.0 * differential (partial) delta calculation, should save performance when recalculating filtered time for tasks

+++ src/objc/IProject.h
@@ -9,7 +9,7 @@
 
 @protocol IProject<NSObject>
 
-- (int) totalTime;
+- (NSInteger) totalTime;
 - (void) updateTotalTime;
 - (NSArray *) tasks;
 - (NSString*) name;
